Ya esta todo separado en una nueva app para el cliente
