+++ clienteApp/templates/publicaciones_list.html
